+++ lib/main.dart
@@ -1,119 +1,96 @@
 import 'package:flutter/material.dart';
-import 'package:font_awesome_flutter/font_awesome_flutter.dart';
-import 'package:google_fonts/google_fonts.dart';
+import './search.dart';
 
 void main() {
-  runApp(MyApp());
+  runApp(const MyApp());
 }
 
 class MyApp extends StatelessWidget {
-  @override
-  Widget build(BuildContext context) {
-    return MaterialApp(
-      title: 'Mapping Matching',
-      theme: ThemeData(
-        textTheme: GoogleFonts.sawarabiMinchoTextTheme(
-          Theme.of(context).textTheme,
-        ),
-        primarySwatch: Colors.blue,
-        appBarTheme: const AppBarTheme(
-          backgroundColor: Colors.white,
-          foregroundColor: Colors.black,
-        ),
-      ),
-      // リスト一覧画面を表示
-      home: SearchPage(),
-    );
-  }
-}
+  const MyApp({Key? key}) : super(key: key);
 
-class SearchPage extends StatefulWidget {
-  @override
-  State<StatefulWidget> createState() {
-    return __State();
-  }
-}
-
-// 検索画面用Widget
-class __State extends State<SearchPage> {
-
+  // This widget is the root of your application.
   @override
   Widget build(BuildContext context) {
     return MaterialApp(
       title: 'Flutter Demo',
-      theme: ThemeData(primarySwatch: Colors.blue),
-      home: Scaffold(
-        body: Center(
-        child: Column(
-          mainAxisSize: MainAxisSize.min,
-          children: <Widget>[
-            SizedBox(
-              child: TextFormField(
-              // initialValue: 'Input text',
-              decoration: InputDecoration(
-                border: OutlineInputBorder(),
-                suffixIcon: Icon(
-                  FontAwesomeIcons.search,
-                ),
-              ),
-            ),
-              height: 50,
-              width: 500,
-            ),
-            Container(
-              padding: EdgeInsets.only(top: 32),
-            ),
-            Row(
-              mainAxisAlignment: MainAxisAlignment.spaceEvenly,
-              children: <Widget>[
-                
-                SizedBox(
-                  child: ElevatedButton.icon(
-                      onPressed: () {},
-                      icon: Icon(FontAwesomeIcons.mapMarkedAlt, size: 45, color: Colors.black),
-                      label: new Text(
-                        '現在地近くから探す',
-                        style: TextStyle(
-                          color: Colors.black,
-                          fontSize: 22,
-                        ),
-                      ),
-                      style: ElevatedButton.styleFrom(
-                        primary: Colors.yellow, //ボタンの背景色
-                      ),
-                    ),
-                  height: 50,
-                  width: 350,
-                ),
-              ],
-            ),
-            Container(
-              padding: EdgeInsets.only(top: 32),
-            ),
-            Row(
-              mainAxisAlignment: MainAxisAlignment.spaceEvenly,
-              children: <Widget>[
+      theme: ThemeData(
+        // This is the theme of your application.
+        //
+        // Try running your application with "flutter run". You'll see the
+        // application has a blue toolbar. Then, without quitting the app, try
+        // changing the primarySwatch below to Colors.green and then invoke
+        // "hot reload" (press "r" in the console where you ran "flutter run",
+        // or simply save your changes to "hot reload" in a Flutter IDE).
+        // Notice that the counter didn't reset back to zero; the application
+        // is not restarted.
+        primarySwatch: Colors.blue,
+      ),
+      home: const MyHomePage(title: 'Flutter Demo Home Page'),
+    );
+  }
+}
 
-                SizedBox(
-                  child: ElevatedButton(
-                  onPressed: () {},
-                  child: new Text(
-                        '節約情報',
-                        style: TextStyle(
-                          color: Colors.black,
-                          fontSize: 25,
-                        ),
-                      ),
-                ),
-                  height: 50,
-                  width: 350,
-                ),
-              ],
-            ),
-          ],
-        ),
+class MyHomePage extends StatefulWidget {
+  const MyHomePage({Key? key, required this.title}) : super(key: key);
+
+  // This widget is the home page of your application. It is stateful, meaning
+  // that it has a State object (defined below) that contains fields that affect
+  // how it looks.
+
+  // This class is the configuration for the state. It holds the values (in this
+  // case the title) provided by the parent (in this case the App widget) and
+  // used by the build method of the State. Fields in a Widget subclass are
+  // always marked "final".
+
+  final String title;
+
+  @override
+  State<MyHomePage> createState() => _MyHomePageState();
+}
+
+class _MyHomePageState extends State<MyHomePage> {
+  int _counter = 0;
+
+  void _incrementCounter() {
+    setState(() {
+      // This call to setState tells the Flutter framework that something has
+      // changed in this State, which causes it to rerun the build method below
+      // so that the display can reflect the updated values. If we changed
+      // _counter without calling setState(), then the build method would not be
+      // called again, and so nothing would appear to happen.
+      _counter++;
+    });
+  }
+
+  @override
+  Widget build(BuildContext context) {
+    // This method is rerun every time setState is called, for instance as done
+    // by the _incrementCounter method above.
+    //
+    // The Flutter framework has been optimized to make rerunning build methods
+    // fast, so that you can just rebuild anything that needs updating rather
+    // than having to individually change instances of widgets.
+    return Scaffold(
+      appBar: AppBar(
+        // Here we take the value from the MyHomePage object that was created by
+        // the App.build method, and use it to set our appbar title.
+        title: Text(widget.title),
+      ),
+      body: Center(
+        // Center is a layout widget. It takes a single child and positions it
+        // in the middle of the parent.
+        child: RaisedButton(
+          child: Text('検索ページへ'),
+          onPressed: () => Navigator.of(context).push(MaterialPageRoute(builder: (context) {
+            return SearchView();
+          })),
         ),
       ),
+      floatingActionButton: FloatingActionButton(
+        onPressed: _incrementCounter,
+        tooltip: 'Increment',
+        child: const Icon(Icons.add),
+      ), // This trailing comma makes auto-formatting nicer for build methods.
     );
   }
 }
